+++ CHANGELOG.md
@@ -36,6 +36,7 @@ This version of the system does not support v13, it is exclusively for Foundry v
 ### Changed
 
 - Refactored CONFIG.statusEffects usage from array to record, in accordance with core migration. (#1166)
+- Compendium abilities now use the name of the associated class's heroic resource instead of a generic "Heroic Resource" if possible. (#1419)
 - Migrated roll modes to the new core message modes. (#1681)
 
 ### Deprecated